refactor: Move phishing and socialengineering into their own folders
